+++ modules/visual-tests/tests/tests.spec.ts
@@ -48,6 +48,7 @@ import { testing as ArticleMeta__test } from '../src/routes/test/ArticleMeta';
 import { testing as Attention__test } from '../src/routes/test/Attention';
 import { testing as BasicTable__test } from '../src/routes/test/BasicTable';
 import { testing as BgBox__test } from '../src/routes/test/BgBox';
+import { testing as Bling__test } from '../src/routes/test/Bling';
 import { testing as BlockQuote__test } from '../src/routes/test/BlockQuote';
 import { testing as BreadCrumbs__test } from '../src/routes/test/BreadCrumbs';
 import { testing as ButtonBack__test } from '../src/routes/test/ButtonBack';
@@ -65,10 +66,12 @@ import { testing as ContentImage__test } from '../src/routes/test/ContentImage';
 import { testing as Datepicker__test } from '../src/routes/test/Datepicker';
 import { testing as ExtraLinks__test } from '../src/routes/test/ExtraLinks';
 import { testing as FeatureList__test } from '../src/routes/test/FeatureList';
+import { testing as FieldGroup__test } from '../src/routes/test/FieldGroup';
 import { testing as FileInput__test } from '../src/routes/test/FileInput';
 import { testing as FooterBadges__test } from '../src/routes/test/FooterBadges';
 import { testing as FooterInfo__test } from '../src/routes/test/FooterInfo';
 import { testing as Footnote__test } from '../src/routes/test/Footnote';
+import { testing as Form__test } from '../src/routes/test/Form';
 import { testing as Gallery__test } from '../src/routes/test/Gallery';
 import { testing as GridBlocks__test } from '../src/routes/test/GridBlocks';
 import { testing as Heading__test } from '../src/routes/test/Heading';
@@ -88,6 +91,7 @@ import { testing as NameCards__test } from '../src/routes/test/NameCards';
 import { testing as NewsHero__test } from '../src/routes/test/NewsHero';
 import { testing as PageFilter__test } from '../src/routes/test/PageFilter';
 import { testing as PageHeading__test } from '../src/routes/test/PageHeading';
+import { testing as Picture__test } from '../src/routes/test/Picture';
 import { testing as ProcessOverview__test } from '../src/routes/test/ProcessOverview';
 import { testing as PullQuote__test } from '../src/routes/test/PullQuote';
 import { testing as RadioButtonsGroup__test } from '../src/routes/test/RadioButtonsGroup';
@@ -95,6 +99,7 @@ import { testing as RadioGroup__test } from '../src/routes/test/RadioGroup';
 import { testing as RelatedLinks__test } from '../src/routes/test/RelatedLinks';
 import { testing as RowBlock__test } from '../src/routes/test/RowBlock';
 import { testing as SearchHeroParagraph__test } from '../src/routes/test/SearchHeroParagraph';
+import { testing as SearchInput__test } from '../src/routes/test/SearchInput';
 import { testing as SearchResults__test } from '../src/routes/test/SearchResults';
 import { testing as Selectbox__test } from '../src/routes/test/Selectbox';
 import { testing as ShareButtons__test } from '../src/routes/test/ShareButtons';
@@ -104,6 +109,7 @@ import { testing as SubHeading__test } from '../src/routes/test/SubHeading';
 import { testing as Tabs__test } from '../src/routes/test/Tabs';
 import { testing as TagPill__test } from '../src/routes/test/TagPill';
 import { testing as TextBlock__test } from '../src/routes/test/TextBlock';
+import { testing as TextButton__test } from '../src/routes/test/TextButton';
 import { testing as TextInput__test } from '../src/routes/test/TextInput';
 import { testing as VSpacer__test } from '../src/routes/test/VSpacer';
 import { testing as WizardLayout__test } from '../src/routes/test/WizardLayout';
@@ -130,6 +136,7 @@ const testingInfos: Record<TestPageLabel, TestingInfo> = {
   Attention: Attention__test,
   BasicTable: BasicTable__test,
   BgBox: BgBox__test,
+  Bling: Bling__test,
   BlockQuote: BlockQuote__test,
   BreadCrumbs: BreadCrumbs__test,
   ButtonBack: ButtonBack__test,
@@ -147,10 +154,12 @@ const testingInfos: Record<TestPageLabel, TestingInfo> = {
   Datepicker: Datepicker__test,
   ExtraLinks: ExtraLinks__test,
   FeatureList: FeatureList__test,
+  FieldGroup: FieldGroup__test,
   FileInput: FileInput__test,
   FooterBadges: FooterBadges__test,
   FooterInfo: FooterInfo__test,
   Footnote: Footnote__test,
+  Form: Form__test,
   Gallery: Gallery__test,
   GridBlocks: GridBlocks__test,
   Heading: Heading__test,
@@ -170,6 +179,7 @@ const testingInfos: Record<TestPageLabel, TestingInfo> = {
   NewsHero: NewsHero__test,
   PageFilter: PageFilter__test,
   PageHeading: PageHeading__test,
+  Picture: Picture__test,
   ProcessOverview: ProcessOverview__test,
   PullQuote: PullQuote__test,
   RadioButtonsGroup: RadioButtonsGroup__test,
@@ -177,6 +187,7 @@ const testingInfos: Record<TestPageLabel, TestingInfo> = {
   RelatedLinks: RelatedLinks__test,
   RowBlock: RowBlock__test,
   SearchHeroParagraph: SearchHeroParagraph__test,
+  SearchInput: SearchInput__test,
   SearchResults: SearchResults__test,
   Selectbox: Selectbox__test,
   ShareButtons: ShareButtons__test,
@@ -186,6 +197,7 @@ const testingInfos: Record<TestPageLabel, TestingInfo> = {
   Tabs: Tabs__test,
   TagPill: TagPill__test,
   TextBlock: TextBlock__test,
+  TextButton: TextButton__test,
   TextInput: TextInput__test,
   VSpacer: VSpacer__test,
   WizardLayout: WizardLayout__test,